Fix Target, Upgrade Extranonce
When a target is created from an U256 we know the U256 beeing le
representaion of target, but when is created from an array we assumes the
array to be be, this can lead to confusion and array should be treated
just as raw data for that reverse() has been removed by From<[u8;32]>
for Target.
Fix a bug in how we compare targets.